ui fixes and removal of dummy data
